refactor(graph-gateway): add query rate limit middleware (I)
